renamed configuration handler to please linter
